Move tests for kron fron sandbox/tests/test_kron.py to theano.tensor.tests.test_slinalg.py
